instafashion/scripts/requirements.txt
absl-py==0.10.0
appnope==0.1.0
argon2-cffi==20.1.0
astroid==2.4.2
astunparse==1.6.3
attrs==20.1.0
autopep8==1.5.4
backcall==0.2.0
bleach==3.1.5
blis==0.4.1
boto==2.49.0
boto3==1.14.43
botocore==1.17.43
CacheControl==0.12.6
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
download==0.3.5
entrypoints==0.3
filelock==3.0.12
firebase-admin==4.3.0
firestore==0.0.8
ftfy==5.8
future==0.18.2
gast==0.3.3
gensim==3.8.3
google-api-core==1.22.0
google-api-python-client==1.10.1
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.1
google-cloud-firestore==1.9.0
google-cloud-storage==1.30.0
google-cloud-vision==1.0.0
google-crc32c==0.1.0
google-pasta==0.2.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
grpcio==1.31.0
h5py==2.10.0
httplib2==0.18.1
idna==2.10
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.12
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.12.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytokenizations==0.6.0
pytz==2020.1
pyvisionproductsearch @ file:///Users/dalequark/codestuffs/pyvisionproductsearch
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
scipy==1.4.1
Send2Trash==1.5.0
sentencepiece==0.1.91
six==1.15.0
smart-open==2.1.0
spacy==2.3.2
spacy-transformers==0.6.2
srsly==1.0.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.4.1
tokenizers==0.5.2
toml==0.10.1
torch==1.6.0
torchcontrib==0.0.2
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
transformers==2.8.0
uritemplate==3.0.1
urllib3==1.25.10
wasabi==0.7.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
